requirements.txt
google-analytics-data>=0.18.0
google-ads>=24.0.0
requests>=2.31.0
facebook-business>=19.0.0
protobuf>=4.25.0
